fix: improve uneven number weapon formatting
Fixes a bug whereby if the number of weapons in a unit was not exactly divisible by the number of models a decimal value would be displayed. This fix doesn't solve the underlying issue as it's non trivial to change the code so that we can break down each weapon set accurately to each model, but it fixes the formatting so that it will display only whole numbers of weapons.